# minor optimizations for deserialization