I added some benchmarks to quantify the win (pretty minor, but a win is a win :p)
Before:
```
test bench_parse_header  ... bench:         145 ns/iter (+/- 1)
test bench_parse_message ... bench:         926 ns/iter (+/- 67)
```
After:
```
test bench_parse_header  ... bench:         117 ns/iter (+/- 0) 
test bench_parse_message ... bench:         816 ns/iter (+/- 11)
```

I've [experimented](https://github.com/little-dude/trust-dns/tree/byteorder) with more ambitious changes but it was not conclusive.