@little-dude 2017-12-06
The travis failure is unrelated:
```
error: component download failed for rustc-x86_64-unknown-linux-gnu

info: caused by: could not download file from 'https://static.rust-lang.org/dist/2017-11-23/rustc-1.22.1-x86_64-unknown-linux-gnu.tar.xz' to '/home/travis/.rustup/downloads/9730e46b2301417a5c5e1410f8ab8b8b6979f448b7fffc5e63221bd1bdaa15a3.partial'

info: caused by: http request returned an unsuccessful status code: 503
```