@bluejekyll 2017-12-06
This Ian really cool! I had no idea there were such savings here!

I’ve actually put trust-dns under the microscope, and am trying to shave off some expensive ops during request etc. I’m mostly focused on the server, these changes will help.

It’s really messy right now, so I don’t have anything to share quite yet. Anyway, this is really awesome. I’ll merge in once we get a clean run etc. I know there are some flaky tests. I’ve been taking some steps at getting them more stable, more work to come obviously.