@little-dude 2017-12-06
I also took a stab at the encoder. I think they will mostly pay off for big
messages (read "bigger than 512 bytes"), but I don't know how to build such
messages so I made up these "no_reservation" benchmarks which are a bit silly.

```
                                            before                     after

test bench_emit_header                      212 ns/iter (+/- 6)        145 ns/iter (+/- 0)
test bench_emit_message                     974 ns/iter (+/- 25)       716 ns/iter (+/- 13)
test bench_emit_message_no_reservation    1,050 ns/iter (+/- 26)       723 ns/iter (+/- 5)
test bench_parse_header                     146 ns/iter (+/- 3)        118 ns/iter (+/- 3)
test bench_parse_header_no_reservation      215 ns/iter (+/- 13)       152 ns/iter (+/- 7)
test bench_parse_message                    944 ns/iter (+/- 74)       802 ns/iter (+/- 10)
```

I'm still making some tests for minor tweaks, but I think I'm mostly done now.